requirements.txt
connexion
swagger-ui-bundle
connexion[swagger-ui]
supabase-py
Flask
python-dotenv
psycopg2-binary